add cast to silent compiler warning with 64bit systems.

+++ lib/ssluse.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: ssluse.c,v 1.237 2009-08-29 04:26:02 gknauf Exp $
+ * $Id: ssluse.c,v 1.238 2009-08-29 04:34:44 gknauf Exp $
  ***************************************************************************/
 
 /*
@@ -1596,7 +1596,7 @@ ossl_connect_step1(struct connectdata *conn,
   }
 
   /* pass the raw socket into the SSL layers */
-  if(!SSL_set_fd(connssl->handle, sockfd)) {
+  if(!SSL_set_fd(connssl->handle, (int)sockfd)) {
      failf(data, "SSL: SSL_set_fd failed: %s",
            ERR_error_string(ERR_get_error(),NULL));
      return CURLE_SSL_CONNECT_ERROR;
